Changed panel heights, removed uneeded events
Previously, tabbed panels could contain panels with different heights.
This required the use of a couple events to update the layout of these
panels when dynamic content was loaded after the panels are created.
Using a fixed height (but resizable) tabbed panel removes the need for
those events and the layout updates.