Fixed problem with page loading
If a tab panel contains multiple panels, one of which is a grid that is
never rendered, the second time that page is loaded causes errors. To
prevent this, items of the tab panel are explicitly removed when the
panel is destroyed and the grid is not loaded until it is rendered.