Update tag store when tags are added or removed
After a tag is added or removed from a user, it is possible that the
list of all the tags may be incorrect. Now, the list is updated after
any change to any user's tags. Also, the StoreManager is used to make
a single tag store available throughout the application.