Added imagettftext patch
PHP 5.3.2 (the version included with RHEL 6) has a bug (see
https://bugs.php.net/bug.php?id=51263) in imagettftext.  The bug results
in rotated text being misaligned and misplaced.  This patch provides an
alternative function that renders the text without any rotation on a
temporary image, rotates that image and then copies that to the final
destination image.  The patch also replaces the use of imagettftext with
this new function in the pChart library.